// apps/api/src/modules/v2-post/application/command/auto-save-article/auto-save-article.command.ts
import { ICommand } from '@nestjs/cqrs';
import { UserDto } from '../../../../v2-user/application/user.dto';
import { MediaDto } from '../../../driving-apdater/dto/request';

export type AutoSaveArticleCommandPayload = {
  id: string;

  actor: UserDto;

  title?: string;

  summary?: string;

  content?: string;

  categories?: string[];

  series?: string[];

  tags?: string[];

  groupIds?: string[];

  coverMedia?: MediaDto;

  wordCount?: number;
};

export class AutoSaveArticleCommand implements ICommand {
  public constructor(public readonly payload: AutoSaveArticleCommandPayload) {}
}